Refactor to super duper kotlinee
